sender/requirements.txt
adrf==0.1.6
asgiref==3.8.1
async-property==0.2.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
h11==0.14.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.30.1
whitenoise==6.7.0
